Module OcplibSimplex

module AssertBounds : sig ... end

module Basic : sig ... end

The main entry point of the library. It provides a functor building each key module of OcplibSimplex.

module Core : sig ... end

module CoreSig : sig ... end

module ExtSigs : sig ... end

module Polys : sig ... end

module Rat2 : sig ... end

A representation of rationals bound, which is a pair of rationals:

module Result : sig ... end

module SolveBounds : sig ... end

module Version : sig ... end