Module Core.Ex

An interface for explanations; in practice, they are labels attached to bounds used for backtracking information on how bounds were discovered. The simplex algorithm does not create explanations: it will only attach empty explanations to bounds, build the union of explanations and print them. It is the user's job to provide the initial explanations when initializing the simplex core.

type t

val empty : t

val union : t -> t -> t

val print : Stdlib.Format.formatter -> t -> unit