Module Core.Var

The type of variables maipulated by the simplex algorithm.

type t

type of variables used in the simplex

val compare : t -> t -> int

compare function on vars

val is_int : t -> bool

is_int v returns true if the variable has integer type, and false otherwise

val print : Stdlib.Format.formatter -> t -> unit

print fmt v prints the given var