Module Basic.Make

Parameters

module Var : ExtSigs.Variables

module R : ExtSigs.Rationals

module Ex : ExtSigs.Explanations

Signature

module Core : CoreSig.S with module Var = Var and type R.t = R.t and type V.t = R.t and module Ex = Ex

The core module defines the different data types used by the project and some functions to handle them.

module Assert : AssertBounds.S with module Core := Core

module Solve : SolveBounds.S with module Core := Core

module Result : Result.S with module Core := Core