Module OcplibSimplex.Rat2

A representation of rationals bound, which is a pair of rationals:

the raw value of the bound

an epsilon used as an offset for representing strict bounds.

For example, for the inequality x < 3, the upper bound of x will be represented by 3 - Ɛ, where Ɛ is a positive integer that will be calculated later.

module type SIG = sig ... end

module Make (R : ExtSigs.Rationals) (V : ExtSigs.Value with type r = R.t) : SIG with module R = R and module V = V