Module SolveBounds.Make

Parameters

module Core : CoreSig.S

Signature

module Core = Core

val solve : Core.t -> Core.t

solve env solves the linear inequalities in the simplex env.

Use Result to retrieve the solution if any.

val maximize : Core.t -> Core.P.t -> Core.t * (Core.P.t * bool) option

maximize env objective finds the maximum value of objective that satisfies the system of linear inequalities in env.

Use Result to retrieve the solution if any.

returns
env,(objective,is_max_bounded)